change package io.xdiamond.common to io.github.xdiamond.common #7
