+++ xdiamond-common/src/main/java/io/github/xdiamond/common/util/ThreadFactoryBuilder.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package io.xdiamond.common.util;
+package io.github.xdiamond.common.util;
 
 import java.lang.Thread.UncaughtExceptionHandler;
 import java.util.concurrent.Executors;
