+++ xdiamond-server/src/main/java/io/github/xdiamond/domain/vo/ResolvedConfig.java
@@ -1,8 +1,8 @@
 package io.github.xdiamond.domain.vo;
 
+import io.github.xdiamond.common.util.Native2ascii;
 import io.github.xdiamond.domain.Config;
 import io.github.xdiamond.domain.Project;
-import io.xdiamond.common.util.Native2ascii;
 
 import java.util.List;
